face detection and recognition/git add/requirement.txt
streamlit
opencv-python
DeepFace
numpy
pillow